+++ src/main/java/com/sergosoft/productservice/service/exception/category/CategoryNotFoundException.java
@@ -4,7 +4,7 @@ public class CategoryNotFoundException extends RuntimeException {
 
     private static final String CATEGORY_NOT_FOUND_MESSAGE = "Category with id %d not found";
 
-    public CategoryNotFoundException(Integer categoryId) {
+    public CategoryNotFoundException(Long categoryId) {
         super(String.format(CATEGORY_NOT_FOUND_MESSAGE, categoryId));
     }
 }
