+++ src/main/java/com/sergosoft/productservice/service/exception/category/ParentCategoryNotFoundException.java
@@ -4,7 +4,7 @@ public class ParentCategoryNotFoundException extends RuntimeException {
 
     private static final String PARENT_CATEGORY_NOT_FOUND_MESSAGE = "Parent category with id %d not found";
 
-    public ParentCategoryNotFoundException(Integer parentId) {
+    public ParentCategoryNotFoundException(Long parentId) {
         super(String.format(PARENT_CATEGORY_NOT_FOUND_MESSAGE, parentId));
     }
 }
